refactor: 리뷰 반영 타입 출처 일관성 및 중복 선언 정리
