+++ src/features/school/hooks/useDocumentEvaluationQueries.ts
@@ -6,7 +6,8 @@ import {
   useCustomSuspenseQuery,
 } from '@/shared/hooks/customQuery'
 import { schoolKeys } from '@/shared/queryKeys'
-import type { PartType, SelectionsSortType } from '@/shared/types/umc'
+import type { PartType } from '@/shared/types/part'
+import type { SelectionsSortType } from '@/shared/types/umc'
 
 import {
   getDocumentEvaluationAnswers,
